Ignore fields with the given field name when hashing.
This is not a permanent solution since it ignores all fields with the name,
regardless of the message they are in. A proper solution needs to either encode
package and nested message names OR use a custom option in the proto spec to
indicate that the field should be skipped.